requirements.txt
watchdog==2.2.0
python-dotenv==0.21.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
Werkzeug==2.0.1
pycryptodome==3.17
